+++ internal/archive/testdata/archive/emmy/cluster.json
