+++ internal/archive/testdata/archive/fritz/cluster.json
